Add custom queue support
- Add an optional queue paramater to broker creation methods: not
  changing the previous behavior, and keeping the api the same
- Add SetBrokerQueue to change queue later on if need be
- Remove hardcoded test checking that the first returned field of the
  redis BLPOP is "celery". According to
  https://redis.io/commands/BLPOP#return-value it will always return the
  name of the queue on which the pop occured. As we only query a single
  list, there is no need to check the name again at this point, the nil
  case is handled before that in both the broker and test code.

Ideally I would have like to be able to simply specify the queue in
Delay() and DelayKwargs() but queues are set via CeleryClient, and
Delay() already has variadic parameters, which would mean either
changin API or creating a new DelayToQueue function.

So for now I went with simply a way to change to a new queue.