worker: release messages without defer
defer causes the function calls to occur at the termination of the
function, rather than at the end of the current block. This causes
defers within a loop to build up a large list of calls in memory and
doesn't run them until the worker terminates, causing a memory leak.

This also fixes a nil pointer dereference when the result is nil - for
example, if runTaskFunc's taskFunc.Call calls a worker function with no
return value.